+++ NEWS.rst
@@ -13,7 +13,10 @@ News
 New Features
 +++++++++++++
 
-* Use duration from `events_find()` as `epochs_end` in `epochs_create()`
+* Use duration from `nk.events_find()` as `epochs_end` in `nk.epochs_create()`
+* Add `nk.find_outliers()` to identify outliers (abnormal values)
+* Add utility function - `nk.check_type()` to return appropriate boolean values of input (integer, list, ndarray, pandas dataframe or pandas series)
+* (experimental) Add error bars in the summary plot method to illustrate standard error of each bin
 
 Fix Bugs
 +++++++++++++